app/requirements.txt
streamlit==1.32.0
plotly==5.18.0
pandas==2.2.0
cassandra-driver==3.28.0 